Added many [1] footnotes to the OAuth2 configuration page:
[1] OAuth2DotPyRunThroug
Jamie Nicolson edited this page on Mar 31, 2017 · 4 revisions
Quick Run-Through with oauth2.py
oauth2.py is a script and library to help developers implement and debug OAuth2 support in Python. Even if you're not programming in Python, the script can be used as a standalone utility for generating and authorizing OAuth tokens, and for generating OAuth2 authentication strings from OAuth tokens.
https://github.com/google/gmail-oauth2-tools/wiki/OAuth2DotPyRunThrough